yanwen: [main-train-srl] reformat; rename to main_train_srl;
